Extract the config file parser
